Last Update in index.html
revision things of all atribattributes :

For video ->
src: Specifies the path to the video file.
controls: Adds video controls, like play, pause, and volume.
autoplay: Automatically starts playing the video when the page loads.
loop: Repeats the video once it ends.
muted: Mutes the video by default.
poster: Specifies an image to be displayed before the video starts playing.
width and height: Specifies the dimensions of the video.

For audio ->
src: Specifies the path to the audio file.
controls: Adds audio controls, like play, pause, and volume.
autoplay: Automatically starts playing the audio when the page loads.
loop: Repeats the audio once it ends.
muted: Mutes the audio by default.
preload: Specifies if and how the audio should be loaded when the page loads ('auto', 'metadata', 'none').

For svg ->
width and height: To set the dimensions.
viewBox: To set the coordinate system.
fill and stroke: To set the colors.

For iframe ->
src: Specifies the URL of the page to embed.
height and width: Define the dimensions.
frameborder: Indicates whether to display a border.
scrolling: Controls the scrollbars.
name: For targeting the iFrame in JavaScript.